Bugfixes and API Improvements
- Bugfix related to WorldGuard and Infusion has been patched [Thanks
given to Simplicitee]
- Small graphical change to Multishot [Thanks given to Simplicitee]

API Overhaul:
There is now a new event you can listen to, the EnchantActivateEvent,
called when a stock enchantment is used. This event should be
immediately added to the code of any custom enchant made with the API,
with a cancel check. This can be used to listen to enchantments being
used or to stop them from being used.